prep splitting client, split models and paginators
