+++ genshin/client/base.py
@@ -1,0 +1,5 @@
+# TODO: everything :(
+#       coming soon
+#
+#
+#                                   I hope
