+++ genshin/models/__init__.py
@@ -1,21 +1,7 @@
 """Pydantic models used by genshin.py"""
-from . import *
-from .abyss import *
-from .activities import *
-from .base import *
-from .calculator import *
-from .character import *
-from .daily import *
-from .diary import *
-from .hoyolab import *
-from .intermap import *
-from .notes import *
-from .stats import *
-from .transaction import *
-from .wish import *
 
-from .honkai_base import *
-from .honkai_character import *
-from .honkai_hoyolab import *
-from .honkai_permanent_modes import *
-from .honkai_stats import *
+from .base import *
+from .daily import *
+from .hoyolab import *
+from .genshin import *
+from .honkai import *
