+++ genshin/models/genshin/__init__.py
@@ -1,0 +1,14 @@
+"""Pydantic models used for the Genshin Impact part of the Hoyolab API."""
+
+from . import *
+from .abyss import *
+from .activities import *
+from .calculator import *
+from .character import *
+from .diary import *
+from .record import *
+from .intermap import *
+from .notes import *
+from .stats import *
+from .transaction import *
+from .wish import *
