+++ genshin/models/honkai/__init__.py
@@ -1,0 +1,7 @@
+"""Pydantic models used for the Honkai Impact part of the Hoyolab API."""
+
+from . import *
+from .character import *
+from .record import *
+from .permanent_modes import *
+from .stats import *
