+++ genshin/paginators/__init__.py
@@ -1,0 +1,5 @@
+"""Paginators used by genshin.py"""
+
+from .base import *
+from .genshin import *
+from .honkai import *
